changed ints to double

+++ src/main.c
@@ -5,9 +5,9 @@
 #include "vadefs.h"
 #include "stdarg.h"
 
-void __point_input(int, int *, int *, int *);
-void __result_callback(int, int, int);
-int * __sattelite_distanses(int);
+void __point_input(int, double *, double *, double *);
+void __result_callback(double, double, double);
+double * __sattelite_distanses(int);
 void __log(char *);
 void __flog(char *, ...);
 
@@ -27,18 +27,18 @@ int main(void) {
     return 0;
 }
 
-void __point_input(int i, int * x, int * y, int * z) {
+void __point_input(int i, double * x, double * y, double * z) {
     printf("Enter %d point coordinates: ", i+1);
-    scanf("%d %d %d", x, y, z);
+    scanf("%f %f %f", x, y, z);
 }
 
 
-void __result_callback(int x, int y, int z) {
-    printf("\nSought point coordinates: x=%d, y=%d, z=%d\n", x, y, z);
+void __result_callback(double x, double y, double z) {
+    printf("\nSought point coordinates: x=%f, y=%f, z=%f\n", x, y, z);
 }
 
 int * __sattelite_distanses(int number) {
-    int * distances = calloc(number, sizeof(int));
+    double * distances = calloc(number, sizeof(double));
     srand(time(NULL));
     for(int i = 0; i < number; i++){
         distances[i] = rand() / MAX_DISTANCE;
